@@ -1,7 +1,7 @@
 class Conversation < ActiveRecord::Base
   
-  belongs_to :sender, :foreign_key => :sender_id, class_name: User
-  belongs_to :recipient, :foreign_key => :recipient_id, class_name: User
+  belongs_to :sender, :foreign_key => :sender_id, class_name: 'User'
+  belongs_to :recipient, :foreign_key => :recipient_id, class_name: 'User'
   
   has_many :messages, dependent: :destroy
   
